+++ CHANGELOG.md
@@ -2,3 +2,16 @@
 
 - Initial version
 - Supports Id3 v1.1 tags
+
+## 0.0.2
+
+- Basic support of ID3 v2.4 tags
+    - support main frames include APIC
+- TagProcessor now can get tags from file and from ByteData
+- You can select which type of tag do you want to get
+- Small fix in id3 v1 parser
+
+## 0.0.3
+
+- Removed dependency from [dart:io]
+- Some API changes